requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pyodbc>=4.0.39
pandas>=2.0.0
python-dotenv>=1.0.0

# Dependências de desenvolvimento/teste (opcional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0